+++ cmd/scam.go
@@ -1,0 +1,23 @@
+package cmd
+
+import (
+	"github.com/kcapp/api/data"
+	"github.com/kcapp/api/models"
+	"github.com/spf13/cobra"
+)
+
+// scamCmd represents the scam command
+var scamCmd = &cobra.Command{
+	Use:   "scam",
+	Short: "Recalculate Scam statistics",
+	Run: func(cmd *cobra.Command, args []string) {
+		err := data.RecalculateStatistics(models.SCAM, legID, since, dryRun)
+		if err != nil {
+			panic(err)
+		}
+	},
+}
+
+func init() {
+	recalculateStatisticsCmd.AddCommand(scamCmd)
+}