+++ data/statistics_420.go
@@ -2,7 +2,7 @@ package data
 
 import (
 	"database/sql"
-	"log"
+	"fmt"
 
 	"github.com/kcapp/api/models"
 )
@@ -392,28 +392,19 @@ func Calculate420Statistics(legID int) (map[int]*models.Statistics420, error) {
 	return statisticsMap, nil
 }
 
-// ReCalculate420Statistics will recaulcate statistics for 420 legs
-func ReCalculate420Statistics() (map[int]map[int]*models.Statistics420, error) {
-	legs, err := GetLegsOfType(models.FOURTWENTY, true)
-	if err != nil {
-		return nil, err
-	}
-
-	s := make(map[int]map[int]*models.Statistics420)
-	for _, leg := range legs {
-		stats, err := Calculate420Statistics(leg.ID)
+// Recalculate420Statistics will recaulcate statistics for 420 legs
+func Recalculate420Statistics(legs []int) ([]string, error) {
+	queries := make([]string, 0)
+	for _, legID := range legs {
+		stats, err := Calculate420Statistics(legID)
 		if err != nil {
 			return nil, err
 		}
 		for playerID, stat := range stats {
-			log.Printf(`UPDATE statistics_420 SET score = %d, total_hit_rate = %f, hit_rate_1 = %f, hit_rate_2 = %f, hit_rate_3 = %f, hit_rate_4 = %f, hit_rate_5 = %f, hit_rate_6 = %f, hit_rate_7 = %f, hit_rate_8 = %f,
-			hit_rate_9 = %f, hit_rate_10 = %f, hit_rate_11 = %f, hit_rate_12 = %f, hit_rate_13 = %f, hit_rate_14 = %f, hit_rate_15 = %f, hit_rate_16 = %f, hit_rate_17 = %f, hit_rate_18 = %f, hit_rate_19 = %f, hit_rate_20 = %f,
-			hit_rate_bull = %f WHERE leg_id = %d AND player_id = %d;`,
+			queries = append(queries, fmt.Sprintf(`UPDATE statistics_420 SET score = %d, total_hit_rate = %f, hit_rate_1 = %f, hit_rate_2 = %f, hit_rate_3 = %f, hit_rate_4 = %f, hit_rate_5 = %f, hit_rate_6 = %f, hit_rate_7 = %f, hit_rate_8 = %f, hit_rate_9 = %f, hit_rate_10 = %f, hit_rate_11 = %f, hit_rate_12 = %f, hit_rate_13 = %f, hit_rate_14 = %f, hit_rate_15 = %f, hit_rate_16 = %f, hit_rate_17 = %f, hit_rate_18 = %f, hit_rate_19 = %f, hit_rate_20 = %f, hit_rate_bull = %f WHERE leg_id = %d AND player_id = %d;`,
 				stat.Score, stat.TotalHitRate, stat.Hitrates[1], stat.Hitrates[2], stat.Hitrates[3], stat.Hitrates[4], stat.Hitrates[5], stat.Hitrates[6], stat.Hitrates[7], stat.Hitrates[8], stat.Hitrates[9], stat.Hitrates[10],
-				stat.Hitrates[11], stat.Hitrates[12], stat.Hitrates[13], stat.Hitrates[14], stat.Hitrates[15], stat.Hitrates[16], stat.Hitrates[17], stat.Hitrates[18], stat.Hitrates[19], stat.Hitrates[20], stat.Hitrates[25], leg.ID, playerID)
+				stat.Hitrates[11], stat.Hitrates[12], stat.Hitrates[13], stat.Hitrates[14], stat.Hitrates[15], stat.Hitrates[16], stat.Hitrates[17], stat.Hitrates[18], stat.Hitrates[19], stat.Hitrates[20], stat.Hitrates[25], legID, playerID))
 		}
-		s[leg.ID] = stats
 	}
-
-	return s, err
+	return queries, nil
 }